requirements.txt
matplotlib==3.9.1
seaborn==0.13.2
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
streamlit==1.36.0
Pillow==10.4.0
